Separate visual and app layer in collections
