+++ src/collections/components/CollectionListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./CollectionListPage";
 export * from "./CollectionListPage";
+export * from "./filters";
