--- src/collections/types.ts
@@ -1,6 +1,0 @@
-import { FilterOpts } from "@saleor/types";
-import { CollectionPublished } from "@saleor/types/globalTypes";
-
-export interface CollectionListFilterOpts {
-  status: FilterOpts<CollectionPublished>;
-}
