+++ src/collections/views/CollectionList/CollectionList.tsx
@@ -49,7 +49,7 @@ import {
   saveFilterTab,
   getFilterQueryParam,
   getFilterOpts
-} from "./filter";
+} from "./filters";
 import { getSortQueryVariables } from "./sort";
 
 interface CollectionListProps {
