+++ src/collections/views/CollectionList/filters.ts
@@ -1,14 +1,13 @@
-import { IntlShape } from "react-intl";
-
 import {
   CollectionFilterInput,
   CollectionPublished
 } from "@saleor/types/globalTypes";
-import { IFilterElement, IFilter } from "@saleor/components/Filter";
+import { IFilterElement } from "@saleor/components/Filter";
 import { maybe, findValueInEnum } from "@saleor/misc";
-import { createOptionsField } from "@saleor/utils/filters/fields";
-import { commonMessages } from "@saleor/intl";
-import { CollectionListFilterOpts } from "../../types";
+import {
+  CollectionListFilterOpts,
+  CollectionFilterKeys
+} from "@saleor/collections/components/CollectionListPage";
 import {
   CollectionListUrlFilters,
   CollectionListUrlFiltersEnum,
@@ -18,13 +17,8 @@ import {
   createFilterTabUtils,
   createFilterUtils
 } from "../../../utils/filters";
-import messages from "./messages";
 
 export const COLLECTION_FILTERS_KEY = "collectionFilters";
-
-export enum CollectionFilterKeys {
-  status = "status"
-}
 
 export function getFilterOpts(
   params: CollectionListUrlFilters
@@ -35,33 +29,6 @@ export function getFilterOpts(
       value: maybe(() => findValueInEnum(status, CollectionPublished))
     }
   };
-}
-
-export function createFilterStructure(
-  intl: IntlShape,
-  opts: CollectionListFilterOpts
-): IFilter<CollectionFilterKeys> {
-  return [
-    {
-      ...createOptionsField(
-        CollectionFilterKeys.status,
-        intl.formatMessage(commonMessages.status),
-        [opts.status.value],
-        false,
-        [
-          {
-            label: intl.formatMessage(messages.published),
-            value: CollectionPublished.PUBLISHED
-          },
-          {
-            label: intl.formatMessage(messages.hidden),
-            value: CollectionPublished.HIDDEN
-          }
-        ]
-      ),
-      active: opts.status.active
-    }
-  ];
 }
 
 export function getFilterVariables(
