--- src/collections/views/CollectionList/messages.ts
@@ -1,14 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  hidden: {
-    defaultMessage: "Hidden",
-    description: "collection"
-  },
-  published: {
-    defaultMessage: "Published",
-    description: "collection"
-  }
-});
-
-export default messages;
